Fix Sonar Issues - [dups]
